# Create optimized TryRead<T>.