--- a/src/System.Buffers.ReaderWriter/System/Buffers/Reader/BufferReader_binary.cs
+++ b/src/System.Buffers.ReaderWriter/System/Buffers/Reader/BufferReader_binary.cs
@@ -1,6 +1,8 @@
 ﻿// Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
+using System.Buffers.Binary;
+using System.Diagnostics;
 using System.Runtime.InteropServices;
 
 namespace System.Buffers.Reader
@@ -29,6 +31,8 @@ namespace System.Buffers.Reader
 
         private unsafe bool TryReadSlow<T>(out T value) where T : unmanaged
         {
+            Debug.Assert(UnreadSegment.Length < sizeof(T));
+
             // Not enough data in the current segment, try to peek for the data we need.
             byte* buffer = stackalloc byte[sizeof(T)];
             Span<byte> tempSpan = new Span<byte>(buffer, sizeof(T));
@@ -42,6 +46,37 @@ namespace System.Buffers.Reader
             value = MemoryMarshal.Read<T>(tempSpan);
             Advance(sizeof(T));
             return true;
+        }
+
+        public unsafe bool TryReadInt32LittleEndian(out int value)
+        {
+            if (BitConverter.IsLittleEndian)
+            {
+                return TryRead(out value);
+            }
+
+            return TryReadReverseEndianness(out value);
+        }
+
+        public unsafe bool TryReadInt32BigEndian(out int value)
+        {
+            if (!BitConverter.IsLittleEndian)
+            {
+                return TryRead(out value);
+            }
+
+            return TryReadReverseEndianness(out value);
+        }
+
+        private unsafe bool TryReadReverseEndianness(out int value)
+        {
+            if (TryRead(out value))
+            {
+                value = BinaryPrimitives.ReverseEndianness(value);
+                return true;
+            }
+
+            return false;
         }
     }
 }