--- a/src/System.Buffers.ReaderWriter/System/Buffers/Reader/BufferReader_text.cs
+++ b/src/System.Buffers.ReaderWriter/System/Buffers/Reader/BufferReader_text.cs
@@ -11,8 +11,12 @@ namespace System.Buffers.Reader
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public unsafe bool TryParse(out bool value)
         {
-            var unread = UnreadSegment;
-            if (Utf8Parser.TryParse(unread, out value, out int consumed) && consumed < unread.Length)
+            ReadOnlySpan<byte> unread = UnreadSegment;
+
+            // For other types (int, etc) we won't know if we've consumed all of the type
+            // ("235612" can be split over segments, for example). For bool, Utf8Parser
+            // doesn't care what follows "True" or "False" and neither should we.
+            if (Utf8Parser.TryParse(unread, out value, out int consumed))
             {
                 Advance(consumed);
                 return true;
@@ -24,7 +28,7 @@ namespace System.Buffers.Reader
         private unsafe bool TryParseSlow(out bool value)
         {
             const int MaxLength = 5;
-            var unread = UnreadSegment;
+            ReadOnlySpan<byte> unread = UnreadSegment;
 
             if (unread.Length > MaxLength)
             {
@@ -47,7 +51,7 @@ namespace System.Buffers.Reader
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public unsafe bool TryParse(out int value)
         {
-            var unread = UnreadSegment;
+            ReadOnlySpan<byte> unread = UnreadSegment;
             if (Utf8Parser.TryParse(unread, out value, out int consumed) && consumed < unread.Length)
             {
                 Advance(consumed);
@@ -60,7 +64,7 @@ namespace System.Buffers.Reader
         private unsafe bool TryParseSlow(out int value)
         {
             const int MaxLength = 15;
-            var unread = UnreadSegment;
+            ReadOnlySpan<byte> unread = UnreadSegment;
 
             if (unread.Length > MaxLength)
             {
@@ -83,7 +87,7 @@ namespace System.Buffers.Reader
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public unsafe bool TryParse(out ulong value)
         {
-            var unread = UnreadSegment;
+            ReadOnlySpan<byte> unread = UnreadSegment;
             if (Utf8Parser.TryParse(unread, out value, out int consumed) && consumed < unread.Length)
             {
                 Advance(consumed);
@@ -96,7 +100,7 @@ namespace System.Buffers.Reader
         private unsafe bool TryParseSlow(out ulong value)
         {
             const int MaxLength = 30;
-            var unread = UnreadSegment;
+            ReadOnlySpan<byte> unread = UnreadSegment;
 
             if (unread.Length > MaxLength)
             {