--- a/tests/Benchmarks/Benchmarks.csproj
+++ b/tests/Benchmarks/Benchmarks.csproj
@@ -17,6 +17,11 @@
     <EmbeddedResource Include="System.Text.Primitives\SampleTexts\*.txt" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="..\System.Buffers.Primitives.Tests\BufferSegment.cs" Link="BufferSegment.cs" />
+    <Compile Include="..\System.Buffers.Primitives.Tests\BufferUtilities.cs" Link="System.Buffers\BufferUtilities.cs" />
+    <Compile Include="..\System.Buffers.Primitives.Tests\SpanLiteralExtensions.cs" Link="SpanLiteralExtensions.cs" />
+  </ItemGroup>
+  <ItemGroup>
     <PackageReference Include="BenchmarkDotNet" Version="$(BenchmarkDotNetVersion)" />
     <PackageReference Include="BenchmarkDotNet.Diagnostics.Windows" Version="$(BenchmarkDotNetVersion)" />
     <PackageReference Include="System.Text.Encodings.Web" Version="$(CoreFxStableVersion)" />