--- a/tests/Benchmarks/System.Buffers/Reader.cs
+++ b/tests/Benchmarks/System.Buffers/Reader.cs
@@ -45,7 +45,7 @@ namespace System.Buffers.Benchmarks
         {
             var reader = BufferReader.Create(s_ros);
 
-            while (BufferReaderExtensions.TryParse(ref reader, out int value))
+            while (reader.TryParse(out int value))
             {
                 reader.Advance(1); // advance past the delimiter
             }