… src/System.Buffers.ReaderWriter/System/Buffers/Transformations/RemoveTransformation.cs
… tests/Benchmarks/System.Buffers/Reader_ParseInt.cs
… tests/System.Buffers.Primitives.Tests/BufferUtilities.cs
… tests/System.Buffers.ReaderWriter.Tests/BufferReaderTests.cs
… tests/System.Buffers.ReaderWriter.Tests/BytesReaderTests.cs
… tests/System.Buffers.ReaderWriter.Tests/ReadableBufferReaderFacts.cs