Move helpers to instance methods, split out slow paths and inline fast paths where needed.

Add some perf tests.

Overhead from implicit struct copies, method calls was significant. The best pattern I can come up with is to keep the fast (single span) reads in tight methods with callouts for aggregation. Changed the pattern of peek to return a direct span rather than a copy when it isn't necessary.

**Before:**

```
    Method |     Mean |    Error |   StdDev |
---------- |---------:|---------:|---------:|
 ReadInt32 | 328.9 us | 2.700 us | 2.525 us |
 PeekSpan  | 2.589 ms | 0.0125 ms | 0.0117 ms |
```

**After**

```
    Method |     Mean |     Error |    StdDev |
---------- |---------:|----------:|----------:|
 ReadInt32 | 102.1 us | 0.8000 us | 0.7483 us |
  PeekSpan | 240.5 us | 2.4859 us | 2.2037 us |
```

Roughly 3x / 10x improvement for binary reads. Parsing reads from 20% (split input) - 37% improvement in the benchmarks I added.