@JeremyKuhne 2018-08-07
I also removed endianness in the binary read. Doesn't seem worth taking a hit for. If we really want it we can add slower overloads.