@davidfowl 2018-08-07
If you're going to pick an endianess to support, you'd want big endian since most networking protocols are big endian.