@benaadams 2018-08-07
> Overhead from implicit struct copies, method calls was significant.

Can you add more detail/examples to "Inlined Struct wrapper .ctors are non-zero cost" https://github.com/dotnet/coreclr/issues/18542