@benaadams 2018-08-07
> If you're going to pick an endianess to support, you'd want big endian since most networking protocols are big endian.

When using can always add a non-branching byteswap after (using `BitConverter.IsLittleEndian`  taken to a `readonly static` to change it to a Jit const); though probably also want @GrabYourPitchforks' "[WIP] Add support for BSWAP intrinsic" https://github.com/dotnet/coreclr/pull/18398

Custom protocols; IPC and file formats aren't necessarily big endian (e.g. why convert to big endian then back to little endian when 90% of the machines are little endian these days)