@benaadams 2018-08-07
Ah, `BitConverter.IsLittleEndian` is an intrinsic now in CoreClr https://github.com/dotnet/coreclr/pull/9789