@JeremyKuhne 2018-08-07
> Can you add more detail/examples to "Inlined Struct wrapper .ctors are non-zero cost" 

@benaadams I didn't look at the resulting IL, I just measured. Here are the numbers I saw (before breaking the slow path out of the main method).

```
Use static method in class that takes instance

    Method |     Mean |    Error |   StdDev |
---------- |---------:|---------:|---------:|
 ReadInt32 | 318.0 us | 5.955 us | 5.279 us |

Use instance method

    Method |     Mean |    Error |   StdDev |
---------- |---------:|---------:|---------:|
 ReadInt32 | 272.9 us | 5.203 us | 5.110 us |
```