@JeremyKuhne 2018-08-07
I've added some endian sample methods `public unsafe bool TryReadInt32LittleEndian(out int value)`. If you're aligned with the proper platform it ends up being just about as fast:

```
                 Method |     Mean |     Error |    StdDev |
----------------------- |---------:|----------:|----------:|
              ReadInt32 | 101.7 us | 0.5264 us | 0.4924 us |
    ReadInt32_BigEndian | 137.2 us | 1.0393 us | 0.9722 us |
 ReadInt32_LittleEndian | 103.5 us | 0.6634 us | 0.6206 us |
```


@mgravell, @davidfowl, @Drawaes, @benaadams Thoughts?