@JeremyKuhne 2018-08-07
Still taking feedback of course- merging in the current state.